[386827] GMF-Tooling to generate web diagrams from the same input artifacts
First commit for a GEF-GWT generator for GMF Tooling GenModel. This generator is based on the gmf-lite generator and runtime,
that is use to produce pure GEF applications. Similarly this generator will produce pure GEF application but meant to be run on GWT
thanks to GEF-GWT (see http://gefgwt.org).

Added bundles:
  - org.eclipse.gmf.codegen.gwt contains diagram related templates
  - org.eclipse.gmf.codegen.gwt contains templates related to the GWT client application
  - org.eclipse.gmf.codegen.gwt.ui contains eclipse UI actions